Major restructuring to disentangle scopescan from other passes and make it more reliable
Instead of DeclarationDescriptors, we now map identifiers to Bindings.
Instead of holding a declared type and a value (which might've had to
be inferred during the scopscan pass), a Binding holds the set of
information *needed* to infer the type later. It also holds more
detailed information about the declaration itself (or import, or
whatever).

This makes scopescan's job much simpler: it just points identifiers at
"things", instead of trying to determine eagerly what those things are.
The rest is left up to typeinfer, which is the consumer of most of that
information anyway.